move category resolution into CategorizedCorpusReader
remove raw
