+++ nltk/corpus/reader/tagged.py
@@ -243,40 +243,14 @@ class CategorizedTaggedCorpusReader(CategorizedCorpusReader, TaggedCorpusReader)
         CategorizedCorpusReader.__init__(self, kwargs)
         TaggedCorpusReader.__init__(self, *args, **kwargs)
 
-    def _resolve(self, fileids, categories):
-        if fileids is not None and categories is not None:
-            raise ValueError("Specify fileids or categories, not both")
-        if categories is not None:
-            return self.fileids(categories)
-        else:
-            return fileids
-
-    def raw(self, fileids=None, categories=None):
-        return TaggedCorpusReader.raw(self, self._resolve(fileids, categories))
-
-    def words(self, fileids=None, categories=None):
-        return TaggedCorpusReader.words(self, self._resolve(fileids, categories))
-
-    def sents(self, fileids=None, categories=None):
-        return TaggedCorpusReader.sents(self, self._resolve(fileids, categories))
-
-    def paras(self, fileids=None, categories=None):
-        return TaggedCorpusReader.paras(self, self._resolve(fileids, categories))
-
     def tagged_words(self, fileids=None, categories=None, tagset=None):
-        return TaggedCorpusReader.tagged_words(
-            self, self._resolve(fileids, categories), tagset
-        )
+        return super().tagged_words(self._resolve(fileids, categories), tagset)
 
     def tagged_sents(self, fileids=None, categories=None, tagset=None):
-        return TaggedCorpusReader.tagged_sents(
-            self, self._resolve(fileids, categories), tagset
-        )
+        return super().tagged_sents(self._resolve(fileids, categories), tagset)
 
     def tagged_paras(self, fileids=None, categories=None, tagset=None):
-        return TaggedCorpusReader.tagged_paras(
-            self, self._resolve(fileids, categories), tagset
-        )
+        return super().tagged_paras(self._resolve(fileids, categories), tagset)
 
 
 class TaggedCorpusView(StreamBackedCorpusView):
